rename util to lib
